pyproject.toml
[project]
name = "over-seas-llm-platform-service"
version = "0.1.0"
description = "多海外LLM统一接入与计费监控系统"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "google-generativeai>=0.8.6",
]
